Add files via upload
This repository is about  building  a linear regression models . Interpretation of the  the output of the models is also provided. This model has been used to compete in the Regression with Abalone Dataset competition (late submission) on Kaggle.com.  The Target label is the number of Rings on the Abalone calculated using Root Mean Squared Logarithmic Error (RMSLE). I have interpreted the linear regression model, provided the code, and investigated assumptions, and interpreted all of my findings.